Det nyeste, nu med pause, og counter til player og enemy planet.

+++ src/Entity/GameObject.java
@@ -7,7 +7,7 @@ import java.awt.Graphics2D;
 public abstract class GameObject{
 	// planets
 	protected int planetSize;
-	protected int planetScoreNumber;
+	protected double planetScoreNumber;
 	protected float xLoc;
 	protected float yLoc;
